Siirrä yleiset funktiot utils-objektiin
