+++ src/main/scala/fi/oph/koski/raportit/YleissivistäväUtils.scala
@@ -1,5 +1,10 @@
 package fi.oph.koski.raportit
 
+import java.time.LocalDate
+import java.time.temporal.ChronoUnit
+
+import fi.oph.koski.raportointikanta._
+import fi.oph.koski.schema.{Aikajakso, Jakso}
 import fi.oph.koski.suoritusote.KoulutusModuuliOrdering.järjestäSuffiksinMukaan
 
 
@@ -101,5 +106,147 @@ object YleissivistäväKurssitOrdering {
     val (koodiarvoCharacters, koodiarvoDigits) = järjestäSuffiksinMukaan(kurssi.koulutusmoduuliKoodiarvo)
 
     (kurssi.koulutusmoduuliPaikallinen, koodiarvoCharacters, koodiarvoDigits)
+  }
+}
+
+trait YleissivistäväRaporttiRows {
+  def opiskeluoikeus: ROpiskeluoikeusRow
+  def henkilo: RHenkilöRow
+  def aikajaksot: Seq[ROpiskeluoikeusAikajaksoRow]
+  def päätasonSuoritus: RPäätasonSuoritusRow
+  def osasuoritukset: Seq[ROsasuoritusRow]
+}
+
+object YleissivistäväUtils {
+  def opetettavatOppiaineetJaNiidenKurssit(
+    isOppiaineenOppimäärä: RPäätasonSuoritusRow => Boolean,
+    isOppiaine: ROsasuoritusRow => Boolean,
+    rows: Seq[YleissivistäväRaporttiRows]
+  ): Seq[YleissivistäväRaporttiOppiaineJaKurssit] = {
+    rows
+      .flatMap(oppianeetJaNiidenKurssit(isOppiaineenOppimäärä, isOppiaine))
+      .groupBy(_.oppiaine)
+      .map { case (oppiaine, x) =>
+      YleissivistäväRaporttiOppiaineJaKurssit(oppiaine, x.flatMap(_.kurssit).distinct.sorted(YleissivistäväKurssitOrdering.yleissivistäväKurssitOrdering))
+    }
+      .toSeq
+      .sorted(YleissivistäväOppiaineetOrdering)
+  }
+
+  private def oppianeetJaNiidenKurssit(
+    isOppiaineenOppimäärä: RPäätasonSuoritusRow => Boolean,
+    isOppiaine: ROsasuoritusRow => Boolean
+  ) (row: YleissivistäväRaporttiRows): Seq[YleissivistäväRaporttiOppiaineJaKurssit] = {
+    if (isOppiaineenOppimäärä(row.päätasonSuoritus)) {
+      Seq(YleissivistäväRaporttiOppiaineJaKurssit(toOppiaine(row.päätasonSuoritus), row.osasuoritukset.map(toKurssi)))
+    } else {
+      oppiaineetJaNiidenKurssitOppimäärästä(isOppiaine, row)
+    }
+  }
+
+  private def oppiaineetJaNiidenKurssitOppimäärästä(
+    isOppiaine: ROsasuoritusRow => Boolean,
+    row: YleissivistäväRaporttiRows
+  ): Seq[YleissivistäväRaporttiOppiaineJaKurssit] = {
+    val kurssit = row.osasuoritukset.filter(_.ylempiOsasuoritusId.isDefined).groupBy(_.ylempiOsasuoritusId.get)
+    val oppiaineet = row.osasuoritukset.filter(isOppiaine)
+    val combineOppiaineWithKurssit = (oppiaine: ROsasuoritusRow) =>
+      YleissivistäväRaporttiOppiaineJaKurssit(
+        toOppiaine(oppiaine),
+        kurssit.getOrElse(oppiaine.osasuoritusId, Nil).map(toKurssi)
+      )
+    oppiaineet.map(combineOppiaineWithKurssit)
+  }
+
+  private def toOppiaine(row: RSuoritusRow) = row match {
+    case s: RPäätasonSuoritusRow =>
+      YleissivistäväRaporttiOppiaine(
+        s.suorituksestaKäytettäväNimi.getOrElse("ei nimeä"),
+        s.koulutusmoduuliKoodiarvo,
+        !s.koulutusmoduuliKoodisto.contains("koskioppiaineetyleissivistava")
+      )
+    case s: ROsasuoritusRow =>
+      YleissivistäväRaporttiOppiaine(
+        s.suorituksestaKäytettäväNimi.getOrElse("ei nimeä"),
+        s.koulutusmoduuliKoodiarvo,
+        s.koulutusmoduuliPaikallinen
+      )
+  }
+
+  private def toKurssi(row: ROsasuoritusRow) = {
+    YleissivistäväRaporttiKurssi(
+      row.koulutusmoduuliNimi.getOrElse("ei nimeä"),
+      row.koulutusmoduuliKoodiarvo,
+      row.koulutusmoduuliPaikallinen
+    )
+  }
+
+  def oppiaineidentiedot(
+    paatasonsuoritus: RPäätasonSuoritusRow,
+    osasuoritukset: Seq[ROsasuoritusRow],
+    oppiaineet: Seq[YleissivistäväRaporttiOppiaineJaKurssit],
+    isOppiaineenOppimäärä: RPäätasonSuoritusRow => Boolean
+  ): Seq[String] = {
+    val oppiaineentiedot = if (isOppiaineenOppimäärä(paatasonsuoritus)) {
+      oppiaineenOppimääränOppiaineenTiedot(paatasonsuoritus, osasuoritukset, oppiaineet)
+    } else {
+      oppimääränOppiaineenTiedot(osasuoritukset, oppiaineet)
+    }
+
+    oppiaineet.map(x => oppiaineentiedot(x.oppiaine).map(_.toString).mkString(","))
+  }
+
+  private def oppiaineenOppimääränOppiaineenTiedot(
+    päätasonSuoritus: RPäätasonSuoritusRow,
+    osasuoritukset: Seq[ROsasuoritusRow],
+    oppiaineet: Seq[YleissivistäväRaporttiOppiaineJaKurssit]
+  ): YleissivistäväRaporttiOppiaine => Seq[YleissivistäväOppiaineenTiedot] = {
+    (oppiaine: YleissivistäväRaporttiOppiaine) => if (päätasonSuoritus.matchesWith(oppiaine)) {
+      Seq(toOppiaineenTiedot(päätasonSuoritus, osasuoritukset))
+    } else {
+      Nil
+    }
+  }
+
+  private def oppimääränOppiaineenTiedot(
+    osasuoritukset: Seq[ROsasuoritusRow],
+    oppiaineet: Seq[YleissivistäväRaporttiOppiaineJaKurssit]
+  ): YleissivistäväRaporttiOppiaine => Seq[YleissivistäväOppiaineenTiedot] = {
+    val osasuorituksetMap = osasuoritukset.groupBy(_.koulutusmoduuliKoodiarvo)
+    val oppiaineenSuoritukset = (oppiaine: YleissivistäväRaporttiOppiaine) => osasuorituksetMap
+      .getOrElse(oppiaine.koulutusmoduuliKoodiarvo, Nil).filter(_.matchesWith(oppiaine))
+    (oppiaine: YleissivistäväRaporttiOppiaine) => oppiaineenSuoritukset(oppiaine)
+      .map(s => YleissivistäväOppiaineenTiedot(
+        s.arviointiArvosanaKoodiarvo,
+        osasuoritukset.count(_.ylempiOsasuoritusId.contains(s.osasuoritusId))
+      ))
+  }
+
+  private def toOppiaineenTiedot(suoritus: RSuoritusRow, osasuoritukset: Seq[ROsasuoritusRow]): YleissivistäväOppiaineenTiedot = {
+    val laajuus = suoritus match {
+      case _: RPäätasonSuoritusRow => osasuoritukset.size
+      case s: ROsasuoritusRow => osasuoritukset.count(_.ylempiOsasuoritusId.contains(s.osasuoritusId))
+    }
+    YleissivistäväOppiaineenTiedot(suoritus.arviointiArvosanaKoodiarvo, laajuus)
+  }
+
+  def removeContinuousSameTila(aikajaksot: Seq[ROpiskeluoikeusAikajaksoRow]): Seq[ROpiskeluoikeusAikajaksoRow] = {
+    if (aikajaksot.size < 2) {
+      aikajaksot
+    } else {
+      val rest = aikajaksot.dropWhile(_.tila == aikajaksot.head.tila)
+      aikajaksot.head +: removeContinuousSameTila(rest)
+    }
+  }
+
+  def lengthInDaysInDateRange(jakso: Jakso, alku: LocalDate, loppu: LocalDate): Int = {
+    val hakuvali = Aikajakso(alku, Some(loppu))
+    if (jakso.overlaps(hakuvali)) {
+      val start = if (jakso.alku.isBefore(alku)) alku else jakso.alku
+      val end = if (jakso.loppu.exists(_.isBefore(loppu))) jakso.loppu.get else loppu
+      ChronoUnit.DAYS.between(start, end).toInt + 1
+    } else {
+      0
+    }
   }
 }
